requirements.txt
Flask
torch
torchvision
Pillow
groq
requests
openai
